surface reconstruction starts after 1 fm/c from the beginning
Initially, for the purpose of investigating the total energy balance I
made a change to start the surface reconstruction (and positive/negative
energy flow calculations through it) at the start of the fluid
evolution.
However, later I reverted the change to the 1 fm/c delay (it
was 2.0 fm/c originally) since in some scenarios the surface was
collapsing before enough hadrons fluidised, which was causing a
premature end of the fluid evolution.